add some test data: gpx and geojson
I went on a biking tour from Klaipeda partway to Nida this summer.
Captured a trace with OSMTracker on Android.  That generated a gpx track file.
Then I used gpsbabel to convert it, like this:

gpsbabel -i gpx -f curonian-spit-biking.gpx -o geojson -F curonian-spit-biking.json

Eventually this viewer ought to be able to display this type of track.